internProj/src/requirements.txt
streamlit
google-generativeai>=0.3.3
python-dotenv
python-dateutil
python-docx
pdfkit
wkhtmltopdf